requirements.txt
Pillow==10.2.0
python-slugify==8.0.4
Flask==3.0.2
Werkzeug==3.0.1
beautifulsoup4==4.12.3 